[CL-1710] Code cleanup and some abstraction
